Add a side=(left|right|both|*) parameter.
A "side hint" may also be specified, in which case that will be used
to determine which side of the way the point is on.

Given the following:

&side=right&loc=wayLocation:sideHint

only an edge which is to the right of "sideHint", and closest to
"wayLocation" will be used for routing.

This is used for creating GTFS shapes for buses: the sideHint-s are the
bus stops (where passengers wait), while the wayLocations are the places
on the road, where the bus actually stops.